Add CC as Recommender and change data type in GetLetterofRecommendationByFormID
(cherry picked from commit 37ded634ceada14014f8d29d9d1292a91d1b4c5f)

+++ ThoughtFocus.Service/Implementation/InitialCredentialProgramService.cs
@@ -1499,7 +1499,7 @@ namespace ThoughtFocus.Service.Implementation
                                                FormID = Convert.ToInt32(row["FormID"]),
                                                RecommenderName = Convert.ToString(row["RecommenderName"]),
                                                RecommenderEmail = Convert.ToString(row["RecommenderEmail"]),
-                                               CreatedBy = Convert.ToInt16(row["CreatedBY"]),
+                                               CreatedBy = Convert.ToInt32(row["CreatedBY"]),
                                                CreatedDate = Convert.ToDateTime(row["CreatedDate"]),
                                                RecommenderURL = Convert.ToString(row["RecommenderURL"]),
                                                RecommenderURLValidTill = Convert.ToDateTime(row["RecommenderURLValidTill"] == DBNull.Value ? null : row["RecommenderURLValidTill"]),
